requirements.txt
torch==2.6.0+cu126
torchaudio==2.6.0+cu126
torchvision==0.21.0+cu126
openai-whisper
numpy
sounddevice
customtkinter
